[dotify] Initial version of dotify
Add a tool (dotify) that can be used to populate the user's directory
with the dot files by symlinking them to the current directory.

Currently, only creates symlinks if the existing target either does not
exist of is a symlink. Optionally, a "force" mode will delete any
existing files or back them up (if the commandline parameter "-b" is
used).