[dotify] Manage files as inclusive list
In order to deal with tricky situations where the source and destination
paths are different (and because the "exclude" list in dotify was
growing quite fast), manage the files to replace in a separate JSON
file.
This file needs to be extended whenever a new dot file is added, but in
the bargain, there's no more accidential symlinks in $HOME and it's
possible to store supplementary data (such as the destination).